Add files via upload
This is what I've got so far. Running the main opens an (ugly) start menu that has options for game mode, map selection, and to start the game. The game opens with the selected map, in a window the size of the selected map. Right now it draws the entire map at once, so we'll need to make changes to allow the sprites to pass under certain layers. Hitting escape while the game is running closes the game, and reopens the menu- so the user can choose a different map to play without closing the program.