Add files via upload
Created Game Window class that builds the screen in which the game runs.

Created Game parent class, the game that will run when the user picks start from the main menu. Each game mode should inherit this class.

The Game class successfully loads a map and an array of bee objects. It has a scrolling camera, allowing for larger maps than the game window. Some weird behavior if the window is bigger than the map that will have to be addressed.